i965/drv: fix exporting buffers with 3 planes and VA_EXPORT_SURFACE_SEPARATE_LAYERS
Imported from intel-vaapi-driver:

To get the plane offset, y_cb_offset or y_cr_offset must be multiplied with
the pitch for the 'Y' plane. See for example i965_DeriveImage().

Without this, exporting formats with 3 planes where cb_cr_pitch == width/2
is broken, because the offset of the third plane is calculated incorrectly.

Authored-By: michaelolbrich