i965/drv: Partially fix mpv.
mpv (likely ffmpeg) for some reason will try to probe all formats using `vaPutImage`, this is dumb.

On older platforms (pre-Kaby Lake) the driver will hitch on P010, which isn't supported on the VPP path.

Instead, force the SW path and make a clear error be printed to inform the user that the application is wrong.